+++ src/sdax/__init__.py
@@ -13,7 +13,16 @@ from .sdax_core import (
     SdaxExecutionError,
     flatten_exceptions,
 )
-from .tasks import AsyncTask, RetryableException, SdaxTaskGroup, TaskFunction, task_func, task_group_task, task_sync_func
+from .tasks import (
+    AsyncTask,
+    RetryableException,
+    SdaxTaskGroup,
+    TaskFunction,
+    join,
+    task_func,
+    task_group_task,
+    task_sync_func,
+)
 
 __version__ = "0.7.0"
 
@@ -22,6 +31,7 @@ __all__ = [
     "RetryableException",
     "SdaxTaskGroup",
     "TaskFunction",
+    "join",
     "task_func",
     "task_group_task",
     "task_sync_func",